+++ src/components/AddTransactionForm/AddTransactionForm.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import styles from './AddTransactionFormNew.module.css';
+import styles from './AddTransactionForm.module.css';
 import FormButton from 'components/common/FormButton/FormButton';
 import icons from '../../images/icons/sprite.svg';
 import { useMediaQuery } from 'react-responsive';
