+++ src/components/ModalEditTransaction/ModalEditTransaction.module.css
@@ -1,101 +1,96 @@
-/* MOBILE */
-.modal-overlay {
+.editModal {
   position: fixed;
-  display: flex;
-  flex-direction: column;
-  top: 0;
+  top: 0px;
   left: 0;
-  width: 100vw;
+
   height: 100vh;
-  align-items: center;
-  z-index: 20;
+  width: 100%;
+
+  z-index: 100;
+  /* opacity: 0;
+  transform: scale(0.5);
+  transition: all 0.35s ease-in-out; */
+
+  overflow: auto;
 }
 
-.modal-overlay::before {
-  background-color: rgba(34, 13, 91, 0.23);
-  content: '';
-  backdrop-filter: blur(3.5px);
-  position: absolute;
-  inset: 0;
-  z-index: -1;
+.editModal.isOpen {
+  opacity: 1;
+  transform: scale(1);
+  transition: all 0.35s ease-in-out;
 }
 
-.modal-content {
+.modalBg {
+  /* background-color: #ffffff1a; */
+
+  background: linear-gradient(
+    0deg,
+    rgba(83, 61, 186, 1) 0%,
+    rgba(80, 48, 154, 1) 43.14%,
+    rgba(106, 70, 165, 1) 73.27%,
+    rgba(133, 93, 175, 1) 120.03%
+  );
+
+  box-shadow: 0px 4px 60px 0px #00000040;
+  backdrop-filter: blur(50px);
+  width: 100%;
+  height: 100%;
+
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 100%;
-  padding: 40px 20px;
-  box-shadow: 0 4px 60px 0 rgba(0, 0, 0, 0.25);
+  justify-content: center;
+
   overflow: hidden;
-  height: 100%;
 }
 
-.modal-content::before {
-  background-color: rgba(255, 255, 255, 0.1);
-  content: '';
-  backdrop-filter: blur(50px);
-  position: absolute;
-  inset: 0;
-  z-index: -1;
-}
+@media screen and (min-width: 768px) {
+  .editModal {
+    /* background-color: hsla(256, 75%, 20%, 0.4); */
+    background-color: rgba(34, 13, 91, 0.231);
 
-@media only screen and (min-width: 768px) {
-  .modal-content {
-    position: absolute;
-    width: 540px;
-    padding: 40px 73px;
-    border-radius: 8px;
-    height: fit-content;
-  }
-}
+    backdrop-filter: blur(3px);
 
-.gradient-background {
-  z-index: -20;
-  position: absolute;
-  width: 454px;
-  height: 454px;
-  background: rgba(47, 21, 176, 0.73);
-  filter: blur(100px);
-  inset: 0;
-}
-
-.close-button {
-  position: absolute;
-  background-color: transparent;
-  top: 10px;
-  right: 10px;
-  border: none;
-  color: transparent;
-  pointer-events: none;
-}
-
-.title {
-  font-weight: 400;
-  font-size: 24px;
-  margin-bottom: 34px;
-  color: white;
-  margin-bottom: 20px;
-}
-
-.close-button:hover {
-  color: darkmagenta;
-}
-
-/* TABLET / DESKTOP */
-@media only screen and (min-width: 768px) {
-  .modal-content {
+    min-width: 100%;
+    min-height: 100vh;
     display: flex;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -60%);
-    border-radius: 8px;
-    width: 540px;
+    justify-content: center;
+    align-items: center;
+
+    padding: 60px;
+    padding-top: 60px;
+
+    /* overflow: auto;
+    padding-top: 60px;
+    padding-bottom: 60px; */
   }
-  .close-button {
-    color: white;
-    cursor: pointer;
-    pointer-events: initial;
+
+  .modalBg {
+    /* width: 540px;
+    height: 514px; */
+
+    width: fit-content;
+    height: fit-content;
+
+    border-radius: 8px;
+    /* background: linear-gradient(270.02deg, #2e1746 3.2%, #2e225f 99.98%); */
+
+    background: linear-gradient(
+      0deg,
+      rgba(83, 61, 186, 0.7) 0%,
+      rgba(80, 48, 154, 0.7) 43.14%,
+      rgba(106, 70, 165, 0.525) 73.27%,
+      rgba(133, 93, 175, 0.133) 120.03%
+    );
+
+    background: linear-gradient(
+      0deg,
+      rgba(83, 61, 186, 1) 0%,
+      rgba(80, 48, 154, 1) 43.14%,
+      rgba(106, 70, 165, 1) 73.27%,
+      rgba(133, 93, 175, 1) 120.03%
+    );
+
+    background: rgba(82, 59, 126, 1);
   }
 }
