--- src/components/ModalEditTransactionNew/ModalEditTransactionNew.module.css
@@ -1,96 +1,0 @@
-.editModal {
-  position: fixed;
-  top: 0px;
-  left: 0;
-
-  height: 100vh;
-  width: 100%;
-
-  z-index: 100;
-  /* opacity: 0;
-  transform: scale(0.5);
-  transition: all 0.35s ease-in-out; */
-
-  overflow: auto;
-}
-
-.editModal.isOpen {
-  opacity: 1;
-  transform: scale(1);
-  transition: all 0.35s ease-in-out;
-}
-
-.modalBg {
-  /* background-color: #ffffff1a; */
-
-  background: linear-gradient(
-    0deg,
-    rgba(83, 61, 186, 1) 0%,
-    rgba(80, 48, 154, 1) 43.14%,
-    rgba(106, 70, 165, 1) 73.27%,
-    rgba(133, 93, 175, 1) 120.03%
-  );
-
-  box-shadow: 0px 4px 60px 0px #00000040;
-  backdrop-filter: blur(50px);
-  width: 100%;
-  height: 100%;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  overflow: hidden;
-}
-
-@media screen and (min-width: 768px) {
-  .editModal {
-    /* background-color: hsla(256, 75%, 20%, 0.4); */
-    background-color: rgba(34, 13, 91, 0.231);
-
-    backdrop-filter: blur(3px);
-
-    min-width: 100%;
-    min-height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    padding: 60px;
-    padding-top: 60px;
-
-    /* overflow: auto;
-    padding-top: 60px;
-    padding-bottom: 60px; */
-  }
-
-  .modalBg {
-    /* width: 540px;
-    height: 514px; */
-
-    width: fit-content;
-    height: fit-content;
-
-    border-radius: 8px;
-    /* background: linear-gradient(270.02deg, #2e1746 3.2%, #2e225f 99.98%); */
-
-    background: linear-gradient(
-      0deg,
-      rgba(83, 61, 186, 0.7) 0%,
-      rgba(80, 48, 154, 0.7) 43.14%,
-      rgba(106, 70, 165, 0.525) 73.27%,
-      rgba(133, 93, 175, 0.133) 120.03%
-    );
-
-    background: linear-gradient(
-      0deg,
-      rgba(83, 61, 186, 1) 0%,
-      rgba(80, 48, 154, 1) 43.14%,
-      rgba(106, 70, 165, 1) 73.27%,
-      rgba(133, 93, 175, 1) 120.03%
-    );
-
-    background: rgba(82, 59, 126, 1);
-  }
-}
