+++ src/components/ModifyTransactionForm/ModifyTransactionForm.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import styles from './ModifyTransactionFormNew.module.css';
+import styles from './ModifyTransactionForm.module.css';
 
 import FormButton from 'components/common/FormButton/FormButton';
 import icons from '../../images/icons/sprite.svg';
