+++ src/components/StatisticsChart/StatisticsChart.jsx
