+++ src/components/StatisticsTable/StatisticsTable.jsx
