+++ src/components/StatisticsTable/StatisticsTable.module.css
