+++ src/index.js
@@ -1,8 +1,8 @@
 // index.js
-import React from 'react';
+
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { BrowserRouter as Router } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { persistor, store } from './redux/store';
 import { PersistGate } from 'redux-persist/integration/react';
 import App from './components/App';
@@ -10,13 +10,11 @@ import './index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <PersistGate loading={null} persistor={persistor}>
-        <Router basename="/BudgetBusters-MoneyGuard">
-          <App />
-        </Router>
-      </PersistGate>
-    </Provider>
-  </React.StrictMode>
+  <Provider store={store}>
+    <PersistGate loading={null} persistor={persistor}>
+      <BrowserRouter basename="/BudgetBusters-MoneyGuard">
+        <App />
+      </BrowserRouter>
+    </PersistGate>
+  </Provider>
 );
